only add regolith if not present

+++ setup-vm.sh
@@ -28,10 +28,12 @@ sudo apt update -qq
 sudo apt full-upgrade --auto-remove -y --purge -qq
 
 # Regolith
-wget -qO - https://regolith-desktop.org/regolith.key | gpg --dearmor | sudo tee /usr/share/keyrings/regolith-archive-keyring.gpg >/dev/null
-echo deb "[arch=amd64 signed-by=/usr/share/keyrings/regolith-archive-keyring.gpg] https://regolith-desktop.org/release-ubuntu-jammy-amd64 jammy main" | sudo tee /etc/apt/sources.list.d/regolith.list
-sudo apt update -qq
-sudo apt install -y -qq regolith-system-ubuntu
+if [ ! -f /etc/apt/sources.list.d/regolith.list ]; then
+    wget -qO - https://regolith-desktop.org/regolith.key | gpg --dearmor | sudo tee /usr/share/keyrings/regolith-archive-keyring.gpg >/dev/null
+    echo deb "[arch=amd64 signed-by=/usr/share/keyrings/regolith-archive-keyring.gpg] https://regolith-desktop.org/release-ubuntu-jammy-amd64 jammy main" | sudo tee /etc/apt/sources.list.d/regolith.list
+    sudo apt update -qq
+    sudo apt install -y -qq regolith-system-ubuntu
+fi
 
 # virtual box dependencies
 sudo apt install -y -qq dkms gcc perl
